added brief testing code“ g

+++ models/rnw.py
@@ -218,7 +218,7 @@ class RNWModel(LightningModule):
             cam_points = self.backproject(depth, inputs["inv_K", 0])
             pix_coords = self.project_3d(cam_points, inputs["K", 0], T)  # [b,h,w,2]
             src_img = self.get_color_input(inputs, frame_id, 0)
-            outputs[("color", frame_id, scale)] = F.grid_sample(src_img, pix_coords, padding_mode="border",
+            outputs[("color", frame_id, scale)] = F.grid_sample(src_img.double(), pix_coords, padding_mode="border",
                                                                 align_corners=False)
         return outputs
